Fix segfaults when running paster
Fixes #2

I nailed down the segmentation faults occurring to using the development
server (Paster, --reload option didn't affect). Of course I don't know
the exact issue with paster but the underlying issue seems to be Alpine
3.6 + some python packages + a too small stack size.

https://github.com/docker-library/python/issues/211

There are several issues across different projects that confirm that
upgrading to Alpine 3.7 solves the issue, so that's what we did.

The upgrade to Alpine 3.7 caused the following issues:

* curl was not part of the image by default, added it to the installed
  packages

* psycopg2 build failed because of:

        https://github.com/psycopg/psycopg2/issues/594
  The fix is to upgrade the psycopg2 version. This is hardcoded in
  CKAN's requirements.txt so we work around it changing the version when
  building the image (this should be upgraded in CKAN core)

Also for good measure we don't use binaries when building the
requirements because of the issues described here:

https://github.com/ckan/ckan/pull/3893