Separate images for production and development
The existing setup works great if you just want to try out CKAN but it's
not really suited for developement or to customize to your needs.

* Although you can install local extensions with DEVELOPMENT_MODE=true,
  as @gerbyzation pointed out in #3 you can't run the tests because you
  are missing the dev requirements and other configurations.
* If you want to customize the build for your instance with your own
  extensions you need to modify the main ckan Dockerfile, which should
  be stable across releases.

I'm still not 100% happy with the approach but I think this is a step on
the right direction. I have separated the images as such:

+-----------+                +----------+
|           |                |          |
| ckan-base +---------------->   ckan   | (production)
|           |                |          |
+-----+-----+                +----------+
      |
      |
+-----v----+                 +----------+
|          |                 |          |
| ckan-dev +----------------->   ckan   | (development)
|          |                 |          |
+----------+                 +----------+

* `ckan-base`: CKAN with all its dependencies, properly configured and
  running on uWSGI (production setup)
* `ckan-dev`: same as the above, plus:
        * Development requirements installed
        * Installs host extensions cloned on the `src` volume
        * Runs on the Paster development server

From these two base images we would build our own customized image
tailored to our project, installing any extensions and requirements
needed.

To use them users will build their custom images in the `ckan` folder,
where there are two Docker files:

* `Dockerfile`: uses `ckan-base`
* `Dockerfile`: uses `ckan-dev`

To use one or the other there are two separate compose files:

* For a production setup: `docker-compose up` (or `docker-compose -f docker-compose.yml up`)
* For a development setup: `docker-compose -f docker-compose.dev.yml up`

There is obviously duplication on the compose files in all the extra
services, but I wasn't successful extending the production compose file
with a service that used the dev Dockerfile.

I guess we could have a `common.yml` file but I think having this clear
separation is worth a bit of duplication.