Implement custom initialization scripts
Fixes #10

Added an entrypoint directory in the base image (created when this one
is extended, as we are using ONBUILD COPY). If children images add sh or
py scripts to this folder, they will get executed at startup time, just
after prerun.py and before starting the web server and supervisor.
Added docs and example.