Ensure a new value for the session key is used on each new container
Empty the value for `beaker.session.secret` in the ini file included in
the image, and set a new random value in the entrypoint script. This
value should probably be managed via env vars (with
`CKAN___BEAKER__SESSION__SECRET`) like the rest of the secrets, but we
provide this as a fallback value.